Fix: Allow Key Bindings with Shift-Modified Keys
This commit allows to set key bindings that contain
a key modified by a Shift key (e.g. `Ctrl + {`).
For example, after pressing `Ctrl + Shift + [`,
a key binding will be set to `Ctrl + {` as opposed
to `Ctrl + Shift + {` as before.

This is achieved by checking whether a key changes its
value when a Shift key is down. If it does, then the
Shift modifier is removed from `mods`. One exception:
if a key binding contains a letter then the Shift
modifier is not removed. This is because, for some
reason, a key value of a letter is never modified by the
Shift modifier and always corresponds to a key value of
a lowercase character. This is already handled in
`terminatorlib/keybindings.py`.

Resolves: #149