Add cell width configuration in preferences
Currently, we have a setting for changing the line height (cell height),
but not for changing cell width (both available in VTE).

Depending on the font used, it is useful to have the ability to set a
little more space between characters.

This commit adds a configuration similar to the existing one for line
height, but for character space (cell width scale [1]).

Thanks Matt Rose for helping and encouraging me to implement this.

1. https://lazka.github.io/pgi-docs/Vte-2.91/classes/Terminal.html#Vte.Terminal.set_cell_width_scale