Add program_name to fix missing icons
By default, GTK sets the prgname to argv[0], which is then used to
choose the correct .desktop file. If the program is run through a
symlink, such as through `/etc/alternatives/x-terminal-emulator`,
argv[0] is not `terminator` so under Wayland the desktop file is not
detected, leading to a missing icon.