Remove all ibus workarounds
I am sick of ibus breaking every single time we issue a new
release, so as of now, I've gone through the code and removed
all of the workarounds that we have put in to try and get
around issues in ibus, and GTK Input Methods in general.

The code as it stands should work, I'm going to mention a
bunch of people in this PR and ask them to beta test this
change.